+++ src/bfabric/entities/order.py
@@ -16,4 +16,4 @@ class Order(Entity):
     def __init__(self, data_dict: dict[str, Any], client: Bfabric | None) -> None:
         super().__init__(data_dict=data_dict, client=client)
 
-    project: HasOne[Project] = HasOne("Project", bfabric_field="project")
+    project: HasOne[Project] = HasOne("Project", bfabric_field="project", optional=True)