+++ docs/changelog.md
@@ -10,6 +10,10 @@ Versioning currently follows `X.Y.Z` where
 
 ## \[Unreleased\]
 
+### Fixed
+
+- `Order.project` is optional
+
 ## \[1.13.15\] - 2025-01-15
 
 ### Added