+++ docs/changelog.md
@@ -10,6 +10,10 @@ Versioning currently follows `X.Y.Z` where
 
 ## \[Unreleased\]
 
+### Added
+
+- `polars_utils.py` which contains functionality to normalize relational fields in tables
+
 ### Fixed
 
 - `Order.project` is optional