+++ docs/changelog.md
@@ -12,6 +12,7 @@ Versioning currently follows `X.Y.Z` where
 
 ### Added
 
+- Add missing `Entity.__contains__` implementation to check if a key is present in an entity.
 - `polars_utils.py` which contains functionality to normalize relational fields in tables
 
 ### Fixed