+++ docs/changelog.md
@@ -10,6 +10,8 @@ Versioning currently follows `X.Y.Z` where
 
 ## \[Unreleased\]
 
+## \[1.13.16\] - 2025-01-22
+
 ### Added
 
 - Add missing `Entity.__contains__` implementation to check if a key is present in an entity.